+++ clients/elrond/errors.go
@@ -11,8 +11,6 @@ var (
 	errInvalidGasValue          = errors.New("invalid gas value")
 	errNilLogger                = errors.New("nil logger")
 	errNilPrivateKey            = errors.New("nil private key")
-	errNilDataGetter     = errors.New("nil elrondClientDataGetter")
-
 	// ErrNoPendingBatchAvailable signals that no pending batch is available
 	ErrNoPendingBatchAvailable = errors.New("no pending batch available")
 )
