+++ clients/elrond/mappers/errors.go
@@ -1,0 +1,8 @@
+package mappers
+
+import "errors"
+
+var (
+	errNilDataGetter = errors.New("nil elrondClientDataGetter")
+	errUnknownToken  = errors.New("unknown token")
+)
